competition_winning_models_malawi/malawi/2nd Place/povert/requirements.txt
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
catboost==0.5.2
lightgbm==2.0.10
numpy==1.14.0
pandas==0.22.0
scikit-learn==0.19.1
scipy==0.19.1
xgboost>=0.6a2
